xrCore: use Abseil for random (both engine and scripts)
I don't like mt19937, I don't like LuaJIT's random. They're predictable
and not-so-fast.
Pull Abseil here (will be useful in future, maybe for hashes and/or flat
maps), but link only the random library. Instead of all-inline
thread-local mt19937, use ConcurrentQueue with absl::BitGen. It's not so
common to generate random numbers on all your cores simultaneously.
Export get_random_{u64,s64,double_below}() to the scripts. With these 3,
you can easily write a good math.random() replacement.
BTW this very pseudo-random32 from COMMON_AI also goes to the trash bin.
It was even used to generate hashes. Use xr_hash there. The only place
where random32 still must be used is trivial_encryptor (PRNG for
encryption and hashing, lol), inline it there.
Despite that there's now one more lib to link (Abs), the actual exe
became smaller. All-inline doesn't always work, esp. thread local and
overloated stuff like mt19937.

+++ ogsr_engine/xrGame/condition_state.h
@@ -18,7 +18,7 @@ public:
 
 protected:
     xr_vector<_world_property> m_conditions;
-    u32 m_hash;
+    u64 m_hash;
 
 public:
     IC CConditionState();
@@ -34,7 +34,7 @@ public:
     IC bool operator<(const CConditionState& condition) const;
     IC CConditionState<_world_property>& operator-=(const CConditionState& condition);
     IC bool operator==(const CConditionState& condition) const;
-    IC u32 hash_value() const;
+    IC u64 hash_value() const;
     IC const _world_property* property(const typename _world_property::condition_type& condition) const;
 };
